feat: merging FPF with ‘slot’ system, cleaning up types
This is a significant refactoring which merges the functionality of frontend-plugin-framework into the slots system. The word “plugin” is no longer used, being fully replaced with “slot” and “widget”.  A plugin is an App, which is expressed as a collection of routes, slots, and widgets.

The slots system has operations that provide the same functionality as FPF but with, hopefully, a simpler and more approachable schema.  All functionality related to overriding the props of a widget has been replaced by:

- A widget “options” operation which allows a widget to get an object of its configuration options when rendering.
- A best practice to pass props from the widget to its rendered HTMLElement child.
- A widget “replace” operation which allows a widget to be replaced with another, enabling it to take advantage of the prop passing best practice.

In this way, the ‘options’ operation is for expected configuration of widgets, whereas replacing a widget is an escape hatch to do anything else.

This commit also straightens out the operation types and helper functions significantly.  I’m hoping this makes it much clearer how to work with them.

An operation has sub-types, such as UI operations, which in turn has widget and layout subtypes.  The configuration for UI operations can be expressed in multiple ways.  Widget operations, for instance, can take one of a few types of “rendering props”, such as a component, an element, an iframe config, or a module federation config.  Some operations take ‘identity props’ to describe the element they create, (an id and optional role), and others take ‘relationship props’ describing another widget related to the operation - such as “insert after”, “replace”, or “remove”.